+++ test/Main.hs
@@ -5,7 +5,9 @@ import qualified Test.TestConvert as Conv
 import qualified Test.TestFuture as Fut
 import qualified Test.TestTime as Tt
 import qualified Test.TestPico as P
+import qualified Test.TestPico2 as P2
 import qualified Test.TestDmyHm as D 
+
 
 main::IO()
 main = do
@@ -15,3 +17,4 @@ main = do
         Tt.main
         P.main
         D.main
+        P2.main