+++ changelog.md
@@ -1,6 +1,12 @@
+#####   1.0.2
+    Tz derive Functor
+    
+    Data.Time.Hora.Future :
+
+        PicoDiff class
+
 #####   1.0.1
     DmyHmp derive Binary
-    
 
 #####   1.0
     initial version